% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BestSegValEXPRMTL.R
\name{BestSegValEXPRMTL}
\alias{BestSegValEXPRMTL}
\title{EXPRMTL Detect best input variables for TreeSegmentation by validation points}
\usage{
BestSegValEXPRMTL(
  chm,
  a,
  b,
  h,
  vp,
  MIN = 0,
  MAX = 1000,
  skipCheck = FALSE,
  filter = 1
)
}
\arguments{
\item{chm}{raster - RasterLayer with Canopy height model}

\item{a}{numeric - function for MovingWindow}

\item{b}{numeric - function for MovingWindow}

\item{h}{numeric - maximum height of Trees}

\item{vp}{Polygon - PointLayer with estimated Positions of Trees.}

\item{MIN}{numeric - minimum area for Crowns. smaller poylgons are cropped}

\item{MAX}{numeric - maximum area for Crowns. larger polygons are cropped}

\item{skipCheck}{development - bolean - if TRUE skips chekcing the inputs}

\item{filter}{numeric values for the moving window, must be odd}
}
\value{
returns a dataframe with several validation scores
}
\description{
Experimantal Version of BestSegVal. Iterates over a,b,h,MIN and supports iterating over filtered chms.uses supervised computed Treepositions to validate best fitting values for a,b and height.
}
\details{
\itemize{
\item if 'skipCheck' = TRUE the input checking is skipped and the itrations starts directly.
\item has implemneted error catching. if an iteration would cause critical stop error, the loops continue. Returns NA for corrupted iterations
\item Input for a,b,h,MIN,filter supports - numeric, single, combination by c() or sequence by seq() to iterate over.
\item Validation Point - supervised computed pointlayer. Use a GIS like Qgis to set points where Trees are estimated.
\item ! even if the function dupport iteration over all parameters (except MAX due to estimated no nneed for MAX Crownareas to clip) it should not be used to iterate over seqences which would be useless.
}
}
\note{
This function is experimental. The 'Brute Force' abbroach to iterate over many parameters may result in very long time to finsih. Preselected samller sample may be more efficent
}
\examples{
# load data
chmpath <-system.file("extdata","exp_chm.tif",package = "CENITH")
chm <- raster::raster(chmpath)
plot(chm)
vppath <-system.file("extdata","exp_vp.shp",package = "CENITH")
vp <- rgdal::readOGR(vppath)
# test new featrues iterating over MIN and filter
x <-BestSegValEXPRMTL(chm,a=0.5,b=0.5,h=0.5,vp,MIN=c(10,30,50),MAX=1000,filter=c(1,3,5))
# view best accuracy (diffenrent iterations can have the same hit rate, but differ in other values)
maxrow <- x[which.max(x$hit),] # search max vale but rturn only 1 value
maxhit <- maxrow$hit
x[which(x$hit==maxhit),]
# test error causing values
test1 <-BestSegValBETA(chm=chm,a=c(0.9,0.91,0.92,0.1),b=0.5,h=1,vp=vp,MIN = 40,MAX = 200,skipCheck = FALSE)
# if the function stops, retry.
}
\author{
Andreas Schönberg
}
