% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CENITH - package.R
\docType{data}
\name{lau_AOI_rgb - data}
\alias{lau_AOI_rgb - data}
\title{An RGB Raster image of the AOI for the tutorial}
\format{
\code{"raster::raster"}
}
\description{
RGB Raster of the AOI in the Lautaret valley in the French Alps. Resolution 0.15 m.
}
