% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CENITH - package.R
\docType{data}
\name{lau_chm3 - data}
\alias{lau_chm3 - data}
\title{A canopy height model (chm)}
\format{
\code{"raster::raster"}
}
\description{
CHM of some trees in side 3/3 in the Lautaret vally in the france alps. Resolution 0.15 meter.
}
