% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CENITH - package.R
\docType{data}
\name{lau_vp - data}
\alias{lau_vp - data}
\title{Supervised computed Treepositions}
\format{
\code{"rgdal::readOGR"}
}
\description{
Estimated position of trees for the first example chm in the Lautaret Valley. It is used to validate the segmentation.
}
