% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CENITH - package.R
\docType{data}
\name{lau_vp_side2 - data}
\alias{lau_vp_side2 - data}
\title{Supervised computed Treepositions}
\format{
\code{"rgdal::readOGR"}
}
\description{
Estimated Position of trees for the second example chm area in the Lautaret Vally. Used to validate Segmentation.
}
