% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CENITH - package.R
\docType{package}
\name{LEGION - package}
\alias{LEGION - package}
\title{CENITH Tree Segmentation Tool}
\description{
In general the \code{CENITH} package is a wrapper for the \code{ForestTools} package (Andrew Plowright and Jean-Romain Roussel,2020).
\code{CENITH} is designed to test for best fitting parameters to performe a Tree Segmentation. Furher it provides a CrossValidation for Tree Segmentation.
}
\details{
CENITH Tree Segmentation Tool
}
\note{
For a more independent and stabel package the source code of some fucntions provided by \code{ForestTools} is used instead of the respectiv function.
Possible changes of \code{ForestTools} in the future will NOT change the workflow of \code{CENITH}
}
\author{
Andreas Schönberg
}
\keyword{package}
