% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPBestSegValEXPRMTL.R
\name{NPBestSegValEXPRMTL}
\alias{NPBestSegValEXPRMTL}
\title{NOPAR EXPRMTL Detect best input variables for TreeSegmentation by validation points}
\usage{
NPBestSegValEXPRMTL(
  chm,
  a,
  b,
  h,
  vp,
  MIN = 0,
  MAX = 1000,
  skipCheck = FALSE,
  filter = NULL
)
}
\arguments{
\item{chm}{raster - RasterLayer with Canopy height model}

\item{a}{numeric - function for MovingWindow}

\item{b}{numeric - function for MovingWindow}

\item{h}{numeric - maximum height of Trees}

\item{vp}{Polygon - PointLayer with estimated Positions of Trees.}

\item{MIN}{numeric - minimum area for Crowns. smaller poylgons are cropped}

\item{MAX}{numeric - maximum area for Crowns. larger polygons are cropped}

\item{skipCheck}{development - bolean - if TRUE skips chekcing the inputs}

\item{filter}{numeric values for the moving window, must be odd}
}
\value{
returns a dataframe with several validation scores
}
\description{
Experimantal Version of BestSegVal. Iterates over a,b,h,MIN and supports iterating over filtered chms.uses supervised computed Treepositions to validate best fitting values for a,b and height.
}
\details{
\itemize{
\item if 'skipCheck' = TRUE the input checking is skipped and the itrations starts directly.
\item has implemneted error catching. if an iteration would cause critical stop error, the loops continue. Returns NA for corrupted iterations
\item Input for a,b,h,MIN,filter supports - numeric, single, combination by c() or sequence by seq() to iterate over.
\item Validation Point - supervised computed pointlayer. Use a GIS like Qgis to set points where Trees are estimated.
\item ! even if the function dupport iteration over all parameters (except MAX due to estimated no nneed for MAX Crownareas to clip) it should not be used to iterate over seqences which would be useless.
}
}
\note{
This function is experimental. The 'Brute Force' abbroach to iterate over many parameters may result in very long time to finsih. Preselected samller sample may be more efficent
}
\examples{
# load data
chmpath <-system.file("extdata","exp_chm.tif",package = "CENITH")
chm <- raster::raster(chmpath)
plot(chm)
vppath <-system.file("extdata","exp_vp.shp",package = "CENITH")
vp <- rgdal::readOGR(vppath)
# lab test if doParalel works
# make sure no open cluster are running
stopCluster()
# 1st test without doPar
start1 <- Sys.time()
np <-NPBestSegValEXPRMTL(chm,a=c(0.1,0.3,0.5),b=c(0.1,0.3,0.5),h=0.5,vp,MIN=c(10,30,50),MAX=1000,filter=c(1,3,5))
stop1 <- Sys.time()
# 2nd test with doPar full cores
start2 <- Sys.time()
fc <-BestSegValEXPRMTL(chm,a=c(0.1,0.3,0.5),b=c(0.1,0.3,0.5),h=0.5,vp,MIN=c(10,30,50),MAX=1000,filter=c(1,3,5),Cores=1)
stop2 <- Sys.time()
# 3rd test with doPar but lesser cores (3)
start3 <- Sys.time()
x1 <-BestSegValEXPRMTL(chm,a=c(0.1,0.3,0.5),b=0.5,h=0.5,vp,MIN=c(10,30,50),MAX=1000,filter=c(1,3,5),Cores=5)
stop3 <- Sys.time()
# check time differences
difftime(start1,stop1) # no Par
difftime(start2,stop2) # full cores
difftime(start3,stop3) # 3 cores (if 8 availibe)
}
\author{
Andreas Schönberg
}
