% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeSeg.R
\name{TreeSeg}
\alias{TreeSeg}
\title{Compute TreeCrown Segments}
\usage{
TreeSeg(chm = NULL, a, b, h, MIN = 0, MAX = 1000)
}
\arguments{
\item{chm}{raster - RasterLayer with Canopy height model}

\item{a}{numeric - function for MovingWindow}

\item{b}{numeric - function for MovingWindow}

\item{h}{numeric - maximum height of Trees}

\item{MIN}{numeric - minimum area for Crowns. smaller poylgons are cropped}

\item{MAX}{numeric - maximum area for Crowns. larger polygons are cropped}
}
\value{
returns a PolygonLayer with segments
}
\description{
computes polygon segments for TreeCrowns
}
\details{
uses a Moving Window x*a+b to detetc local Maxima in a chm to compute TreeCrown Segments
\itemize{
\item parameter selection - use BestSegVal to automated detect best fitting parameters for a,b,h.
}
}
\note{

}
\examples{
# load data
chmpath <-system.file("extdata","exp_chm.tif",package = "CENITH")
chm <- raster::raster(chmpath)
plot(chm)
# start segmentation
x <- TreeSeg(chm,0.5,0.5,13)
plot(x)
}
\references{

}
\author{

}
