% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeSegCV.R
\name{TreeSegCV}
\alias{TreeSegCV}
\title{TreeSegCV}
\usage{
TreeSegCV(sides, a, b, h, MIN, MAX, vps)
}
\arguments{
\item{sides}{list - chm rasterlayer (see details)}

\item{a}{numeric - function for MovingWindow}

\item{b}{numeric - function for MovingWindow}

\item{h}{numeric - maximum height of Trees}

\item{MIN}{numeric - minimum area for Crowns. smaller poylgons are cropped}

\item{MAX}{numeric - maximum area for Crowns. larger polygons are cropped}

\item{vps}{list - of PointLayers with estimated Positions of Trees (see details)}

\item{skipCheck}{development - bolean - if TRUE skips chekcing the inputs}
}
\value{
returns a dataframe with quality values for the folds and the mean values for a estimted mean quality
}
\description{
cs
}
\details{
*'sides' and 'vps' must be lists with same order of chm and respectiv validtion points
\itemize{
\item a,b,h,MIN,MAX values should be calculated by 'TreeSegVal'
}
}
\author{
Andreas Schönberg
}
