% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BestSegVal.R
\name{BestSegVal}
\alias{BestSegVal}
\title{Detect best input variables for TreeSegmentation by validation points}
\usage{
BestSegVal(chm, a, b, h, vp)
}
\arguments{
\item{chm}{raster - RasterLayer with Canopy height model}

\item{a}{numeric - function for MovingWindow}

\item{b}{numeric - function for MovingWindow}

\item{h}{numeric - maximum height of Trees}

\item{vp}{Polygon - PointLayer with estimated Positions of Trees.}
}
\value{
returns a dataframe with several validation scores
}
\description{
uses supervised computed Treepositions to validate best fitting values for a,b and height.
}
\details{
\itemize{
\item Input for a,b,h supports - numeric, single, combination by c() or sequence by seq() to iterate over.
\item Validation Point - supervised computed pointlayer. Use a GIS like Qgis to set points where Trees are estimated.
}
}
\examples{
# load data
chmpath <-system.file("extdata","exp_chm.tif",package = "CENITH")
chm <- raster::raster(chmpath)
plot(chm)
vppath <-system.file("extdata","exp_vp.shp",package = "CENITH")
vp <- rgdal::readOGR(vppath)
# run validation
x <-BestSegVal(chm,seq(0.7,0.8,0.1),seq(0.5,0.7,0.1),c(13,8),vp)
# view best accuracy (diffenrent iterations can have the same hit rate, but differ in other values)
maxrow <- x[which.max(x$hit),] # search max vale but rturn only 1 value
maxhit <- maxrow$hit
x[which(x$hit==maxhit),]
}
\author{
Andreas Schönberg
}
