% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CENITH - package.R
\docType{data}
\name{lau_AOI_chm - data}
\alias{lau_AOI_chm - data}
\title{A canopy height model (chm) of the AOI for the tutorial}
\format{
\code{"raster::raster"}
}
\description{
chm of the AOI in the Lautaret valley in the French Alps. Resolution 0.15 m.
}
