% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CENITH - package.R
\docType{data}
\name{lau_AOI_rgb - data}
\alias{lau_AOI_rgb - data}
\title{An rgb image for the tutorial AOI}
\format{
\code{"raster::raster"}
}
\description{
AOI rgb of Tutorial in the Lautaret vally in the france alps. Resolution 0.15 meter.
}
