% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CENITH - package.R
\docType{data}
\name{lau_chm3 - data}
\alias{lau_chm3 - data}
\title{A canopy height model (chm)}
\format{
\code{"raster::raster"}
}
\description{
CHM of trees in site 3 of 3 in the Lautaret valley in the French Alps. Resolution 0.15 meter.
}
