% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeSegCV.R
\name{TreeSegCV}
\alias{TreeSegCV}
\title{TreeSegCV}
\usage{
TreeSegCV(sides, a, b, h, MIN, MAX, CHMfilter, vps)
}
\arguments{
\item{sides}{list - chm rasterlayer (see details)}

\item{a}{numeric - function for MovingWindow}

\item{b}{numeric - function for MovingWindow}

\item{h}{numeric - maximum height of Trees}

\item{MIN}{numeric - minimum area for Crowns. smaller poylgons are cropped}

\item{MAX}{numeric - maximum area for Crowns. larger polygons are cropped}

\item{vps}{list - of PointLayers with estimated Positions of Trees (see details)}

\item{skipCheck}{development - bolean - if TRUE skips checking the inputs}
}
\value{
returns a table with quality values for each fold (side) and calculated overall performance (mean values) for all sides.
}
\description{
performs a n-fold cross validation to estimate the performance of a segmentation 'model' for an AOI
}
\details{
*'sides' and 'vps' must be lists with same order of chm and respective validation points.
\itemize{
\item parameters of the model: a,b,h,MIN,MAX should be calculated by 'TreeSegVal'
\item The overall performance is the mean of all sides.
\item For the details of the table values see 'BestSegVal'.
}
}
\note{
The overall performance helps to estimate a precision for an AOI but does NOT give the "exact" precision. More folds will increase the
expressiveness but will need more time to set Validation Points. Further the supervised setting of Validation Points are highly subjectiv and must not mean the real amount and or position of trees.
}
\examples{
require(CENITH)
require(mapview)
require(raster)
require(rgdal)
# load data
chmpath  <-system.file("extdata","lau_chm.tif",package = "CENITH")
chmpath2 <-system.file("extdata","lau_chm_side2.tif",package = "CENITH")
chmpath3 <-system.file("extdata","lau_chm_side3.tif",package = "CENITH")
vppath <-system.file("extdata","lau_vp.shp",package = "CENITH")
vppath2 <-system.file("extdata","lau_vp_side2.shp",package = "CENITH")
vppath3 <-system.file("extdata","lau_vp_side3.shp",package = "CENITH")

chm  <- raster::raster(chmpath)
chm2 <- raster::raster(chmpath2)
chm3 <- raster::raster(chmpath3)
vp <- rgdal::readOGR(vppath)
vp2 <- rgdal::readOGR(vppath2)
vp3 <- rgdal::readOGR(vppath3)

# list all sides and validation points
chmlist <- list(chm,chm2,chm3)
vplist <- list(vp,vp2,vp3)

# run 3 fold cross validation with parameters computed by 'BestSegVal' (from example)
cv <- CENITH::TreeSegCV(sides=chmlist,a=0.3,b=0.5,h=0.5,MIN=5,MAX=1000,CHMfilter=3,vps=vplist)
cv
### the model trained with BestSegVal on side1 reaches a overall performance of 0.77 @ 0.12 for all tree sides.
### Note that the performance on sides 2 and 3 is even better than on side 1 where it was trained. This effect is caused probably because sides 2 and 3 are more homogenious.
### For sure this is just an example and it can happen that the performance on one side is worst. The overall mean performance is used to estimate the quality of an segmentation for an AOI by testing some subareas.
}
\author{
Andreas Schönberg
}
